chore: Move all the Pages to a separate folder
