+++ src/components/index.ts
@@ -1,19 +1,19 @@
 export * from './AddComment'
 export * from './AddFeedback'
-export * from './AddFeedbackPage'
+export * from '../pages/AddFeedbackPage'
 export * from './Card'
 export * from './CattegoryButton'
 export * from './Comment'
 export * from './EditFeedback'
-export * from './EditFeedbackPage'
+export * from '../pages/EditFeedbackPage'
 export * from './FeedBackCard'
-export * from './FeedbackPage'
+export * from '../pages/FeedbackPage'
 export * from './Header'
 export * from './HeaderBox'
-export * from './LoginPage'
-export * from './MainPage'
+export * from '../pages/LoginPage'
+export * from '../pages/MainPage'
 export * from './ManageBar'
 export * from './MenuDropdown'
 export * from './NoFeedback'
-export * from './RoadmapPage'
+export * from '../pages/RoadmapPage'
 export * from './SortByMenu'
