+++ src/pages/MainPage.tsx
@@ -1,9 +1,8 @@
 import { useState } from 'react'
 import { Link } from 'react-router-dom'
-import { Card } from './Card'
-import { Category, Header } from './Header'
-import ManageBar from './ManageBar'
-import NoFeedback from './NoFeedback'
+import { Card, Category, Header } from '../components'
+import ManageBar from '../components/ManageBar'
+import NoFeedback from '../components/NoFeedback'
 
 type Feedback = {
   id: number
